OOP and polymorphism for parameter groups, not complete
